+++ src/main/java/org/example/components/empolyee/EmployeeService.java
@@ -1,6 +1,8 @@
 package org.example.components.empolyee;
 
+import liquibase.pro.packaged.R;
 import org.example.auth.UserService;
+import org.example.components.address.AddressRepository;
 import org.example.components.shop.ShopRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,16 +20,19 @@ public class EmployeeService {
     private final MyEmployeeRepository myEmployeeRepository;
     private final ShopRepository shopRepository;
     private final UserService userService;
+    private final AddressRepository addressRepository;
     @Autowired
     public EmployeeService(EmployeeRepository employeeRepository, MyEmployeeRepository myEmployeeRepository,
-                           UserService userService, ShopRepository shopRepository) {
+                           UserService userService, ShopRepository shopRepository,
+                           AddressRepository addressRepository) {
         this.employeeRepository = employeeRepository;
         this.myEmployeeRepository = myEmployeeRepository;
         this.userService = userService;
         this.shopRepository = shopRepository;
+        this.addressRepository = addressRepository;
     }
 
-    public ResponseEntity<String> findAll(HttpServletRequest request) {
+    public ResponseEntity<?> findAll(HttpServletRequest request) {
         if (!userService.checkAdmin(request)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
         }
@@ -37,11 +42,14 @@ public class EmployeeService {
             employees2.add(EmployeeDTO.fromEmployee(employee));
         }
 
-        return ResponseEntity.ok(employees2.toString());
+        return ResponseEntity.ok(employees2);
     }
     public ResponseEntity<String> save(Employee employee, HttpServletRequest request, Long shop_id){
         if (!userService.checkAdmin(request)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
+        }
+        if (shopRepository.getById(shop_id) == null){
+            return new ResponseEntity<>("Shop not found", HttpStatus.NOT_FOUND);
         }
         employee.setShop(shopRepository.getById(shop_id));
         employeeRepository.save(employee);
@@ -52,11 +60,14 @@ public class EmployeeService {
         if (!userService.checkAdmin(request)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
         }
+        if (employeeRepository.getById(id) == null){
+            return new ResponseEntity<>("Entity not found", HttpStatus.NOT_FOUND);
+        }
         employeeRepository.deleteById(id);
         return ResponseEntity.ok("Deleted");
     }
 
-    public ResponseEntity<String> filter(String name, String email, String phone, Double salary, String position,
+    public ResponseEntity<?> filter(String name, String email, String phone, Double salary, String position,
                                          HttpServletRequest request){
         if (!userService.checkAdmin(request)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
@@ -66,18 +77,42 @@ public class EmployeeService {
         for (Employee employee : employees){
             employees2.add(EmployeeDTO.fromEmployee(employee));
         }
-        return ResponseEntity.ok(employees2.toString());
+        return ResponseEntity.ok(employees2);
     }
 
-    public ResponseEntity<String> findById(Long id, HttpServletRequest request){
+    public ResponseEntity<?> findById(Long id, HttpServletRequest request){
         if (!userService.checkAdmin(request)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
         }
         Optional<Employee> employee = employeeRepository.findById(id);
         if (!employee.isPresent()){
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Entity not found");
         }else{
-            return ResponseEntity.ok((EmployeeDTO.fromEmployee(employee.get())).toString());
+            return ResponseEntity.ok((EmployeeDTO.fromEmployee(employee.get())));
         }
+    }
+
+    public ResponseEntity<String> update(Employee employee, Long id, Long address_id, Long shop_id,
+                                         HttpServletRequest request){
+        if (!userService.checkAdmin(request)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Access denied");
+        }
+        Optional<Employee> employee1 = employeeRepository.findById(id);
+        if (!employee1.isPresent()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Entity not found");
+        }
+        if (address_id != null){
+            employee.setAddress(addressRepository.getById(id));
+        }else{
+            employee.setAddress(employee1.get().getAddress());
+        }
+        if (shop_id != null){
+            employee.setShop(shopRepository.getById(id));
+        }else{
+            employee.setShop(employee1.get().getShop());
+        }
+        employee.setId(id);
+        employeeRepository.save(employee);
+        return new ResponseEntity<>("Updated", HttpStatus.OK);
     }
 }
